``1.5.0`` Created elog.py and added fixes to browser.py
Added:
  +  elog.py

Added some fixes to the TemporaryDirectory class.

Added some fixes to browser.py
  -  Fixed dump_files function
  -  Automatically removes dump files

elog is a file that is basically a todo list.
elog stands for executed log. It is used to keep track of things you need to execute in a file.

This type of file would be used to keep track of progress in a program that may take some time to execute.

If you exit the program you have to start all over, but if you keep track of all that you have executed. You can start from what you were in the progress of doing.